+++ app/mailboxes/mailbox_helper.rb
@@ -112,11 +112,6 @@ module MailboxHelper
     Rails.logger.info "[MailboxHelper] Contact created with ID: #{@contact.id} for inbox with ID: #{@inbox.id}"
   end
 
-  def notification_email_from_chatwoot?
-    # notification emails are send via mailer sender email address. so it should match
-    @processed_mail.original_sender == Mail::Address.new(ENV.fetch('MAILER_SENDER_EMAIL', 'Chatwoot <accounts@chatwoot.com>')).address
-  end
-
   def mail_content
     if processed_mail.text_content.present?
       processed_mail.text_content[:reply]
